Add retries to SAR service calls on throttle
