+++ tests/plugins/application/test_serverless_app_plugin.py
@@ -220,10 +220,120 @@ class TestServerlessAppPlugin_on_before_transform_template_translate(TestCase):
 
         self.assertEqual("value1", output)
 
+    @patch("samtranslator.plugins.application.serverless_app_plugin.SamTemplate")
+    def test_sar_throttling_doesnt_stop_processing(self, SamTemplateMock):
+        client = Mock()
+        client.create_cloud_formation_template = Mock()
+        client.create_cloud_formation_template.side_effect = ClientError(
+            {"Error": {"Code": "TooManyRequestsException"}}, "CreateCloudFormationTemplate"
+        )
+
+        app_resources = [
+            ("id1", ApplicationResource(app_id="id1", semver="1.0.0", location=True)),
+        ]
+
+        sam_template = Mock()
+        SamTemplateMock.return_value = sam_template
+        sam_template.iterate = Mock()
+        sam_template.iterate.return_value = app_resources
+
+        self.plugin = ServerlessAppPlugin(sar_client=client)
+        self.plugin._can_process_application = Mock()
+        self.plugin._can_process_application.return_value = True
+        self.plugin._get_sleep_time_sec = Mock()
+        self.plugin._get_sleep_time_sec.return_value = 0.02
+        self.plugin.TEMPLATE_WAIT_TIMEOUT_SECONDS = 1.0
+
+        self.plugin.on_before_transform_template({})
+        self.assertEqual(
+            self.plugin._applications.get(("id1", "1.0.0")).message,
+            "Resource with id [id1] is invalid. Failed to call SAR, timeout limit exceeded.",
+        )
+        # confirm we had at least two attempts to call SAR and that we executed a sleep
+        self.assertGreater(client.create_cloud_formation_template.call_count, 1)
+        self.assertGreaterEqual(self.plugin._get_sleep_time_sec.call_count, 1)
+
+    @patch("samtranslator.plugins.application.serverless_app_plugin.SamTemplate")
+    def test_unexpected_sar_error_stops_processing(self, SamTemplateMock):
+        template_dict = {"a": "b"}
+        app_resources = [
+            ("id1", ApplicationResource(app_id="id1", semver="1.0.0", location=True)),
+        ]
+
+        sam_template = Mock()
+        SamTemplateMock.return_value = sam_template
+        sam_template.iterate = Mock()
+        sam_template.iterate.return_value = app_resources
+
+        client = Mock()
+        client.create_cloud_formation_template.side_effect = ClientError(
+            {"Error": {"Code": "BadBadError"}}, "CreateCloudFormationTemplate"
+        )
+        self.plugin = ServerlessAppPlugin(sar_client=client)
+        self.plugin._can_process_application = Mock()
+        self.plugin._can_process_application.return_value = True
+
+        with self.assertRaises(ClientError):
+            self.plugin.on_before_transform_template(template_dict)
+
+    @patch("samtranslator.plugins.application.serverless_app_plugin.SamTemplate")
+    def test_sar_success_one_app(self, SamTemplateMock):
+        template_dict = {"a": "b"}
+        app_resources = [
+            ("id1", ApplicationResource(app_id="id1", semver="1.0.0", location=True)),
+        ]
+
+        sam_template = Mock()
+        SamTemplateMock.return_value = sam_template
+        sam_template.iterate = Mock()
+        sam_template.iterate.return_value = app_resources
+
+        client = Mock()
+        client.create_cloud_formation_template = Mock()
+        client.create_cloud_formation_template.return_value = {"TemplateUrl": "/URL", "Status": STATUS_ACTIVE}
+        self.plugin = ServerlessAppPlugin(sar_client=client)
+        self.plugin._can_process_application = Mock()
+        self.plugin._can_process_application.return_value = True
+        self.plugin.on_before_transform_template(template_dict)
+
+        self.assertEqual(client.create_cloud_formation_template.call_count, 1)
+
+    @patch("samtranslator.plugins.application.serverless_app_plugin.SamTemplate")
+    def test_sleep_between_sar_checks(self, SamTemplateMock):
+        template_dict = {"a": "b"}
+        client = Mock()
+
+        app_resources = [
+            ("id1", ApplicationResource(app_id="id1", semver="1.0.0", location=True)),
+        ]
+
+        sam_template = Mock()
+        SamTemplateMock.return_value = sam_template
+        sam_template.iterate = Mock()
+        sam_template.iterate.return_value = app_resources
+        client.create_cloud_formation_template = Mock()
+        client.create_cloud_formation_template.side_effect = [
+            ClientError({"Error": {"Code": "TooManyRequestsException"}}, "CreateCloudFormationTemplate"),
+            {"TemplateUrl": "/URL", "Status": STATUS_ACTIVE},
+        ]
+        self.plugin._can_process_application = Mock()
+        self.plugin._can_process_application.return_value = True
+        self.plugin = ServerlessAppPlugin(sar_client=client, wait_for_template_active_status=True, validate_only=False)
+        self.plugin._get_sleep_time_sec = Mock()
+        self.plugin._get_sleep_time_sec.return_value = 0.001
+        self.plugin.on_before_transform_template(template_dict)
+        # should have exactly two calls to SAR
+        self.assertEqual(client.create_cloud_formation_template.call_count, 2)
+        self.assertEqual(self.plugin._get_sleep_time_sec.call_count, 1)  # make sure we slept once
+
 
 class ApplicationResource(object):
-    def __init__(self, app_id="app_id", semver="1.3.5"):
-        self.properties = {"ApplicationId": app_id, "SemanticVersion": semver}
+    def __init__(self, app_id="app_id", semver="1.3.5", location=None):
+        self.properties = (
+            {"ApplicationId": app_id, "SemanticVersion": semver}
+            if not location
+            else {"Location": {"ApplicationId": app_id, "SemanticVersion": semver}}
+        )
 
 
 # class TestServerlessAppPlugin_on_before_transform_resource(TestCase):
